Cleanup

+++ kernel/syscall.c
@@ -1,8 +1,5 @@
 #include "types.h"
 #include "defs.h"
-#include "param.h"
-#include "memlayout.h"
-#include "mmu.h"
 #include "proc.h"
 #include "x86.h"
 #include "syscall.h"
@@ -20,13 +17,13 @@
  * @param ip Destination pointer in kernel space.
  * @return 0 on success, -1 if the address is invalid.
  */
-int fetchint(u32 addr, int* ip)
+int fetchint(u32 addr, int *ip)
 {
-    struct proc* curproc = myproc();
+    struct proc *curproc = myproc();
 
     if (addr >= curproc->size || addr + 4 > curproc->size)
         return -1;
-    *ip = *(int*)(addr);
+    *ip = *(int *)(addr);
     return 0;
 }
 
@@ -37,16 +34,15 @@ int fetchint(u32 addr, int* ip)
  * @param pp Receives the user pointer.
  * @return Length excluding terminator, or -1 if invalid.
  */
-int fetchstr(u32 addr, char** pp)
+int fetchstr(u32 addr, char **pp)
 {
-    struct proc* curproc = myproc();
+    struct proc *curproc = myproc();
 
     if (addr >= curproc->size)
         return -1;
-    *pp = (char*)addr;
-    char* ep = (char*)curproc->size;
-    for (char* s = *pp; s < ep; s++)
-    {
+    *pp      = (char *)addr;
+    char *ep = (char *)curproc->size;
+    for (char *s = *pp; s < ep; s++) {
         if (*s == 0)
             return s - *pp;
     }
@@ -60,7 +56,7 @@ int fetchstr(u32 addr, char** pp)
  * @param ip Destination pointer.
  * @return 0 on success, -1 on failure.
  */
-int argint(int n, int* ip)
+int argint(int n, int *ip)
 {
     return fetchint((myproc()->trap_frame->esp) + 4 + 4 * n, ip);
 }
@@ -73,16 +69,16 @@ int argint(int n, int* ip)
  * @param size Size in bytes that must fit within process memory.
  * @return 0 on success, -1 on failure.
  */
-int argptr(int n, char** pp, int size)
+int argptr(int n, char **pp, int size)
 {
     int i;
-    struct proc* curproc = myproc();
+    struct proc *curproc = myproc();
 
     if (argint(n, &i) < 0)
         return -1;
     if (size < 0 || (u32)i >= curproc->size || (u32)i + size > curproc->size)
         return -1;
-    *pp = (char*)i;
+    *pp = (char *)i;
     return 0;
 }
 
@@ -93,7 +89,7 @@ int argptr(int n, char** pp, int size)
  * @param pp Receives the user string pointer.
  * @return 0 on success, -1 on failure.
  */
-int argstr(int n, char** pp)
+int argstr(int n, char **pp)
 {
     int addr;
     if (argint(n, &addr) < 0)
@@ -153,17 +149,16 @@ static int (*syscalls[])(void) = {
  */
 void syscall(void)
 {
-    struct proc* curproc = myproc();
+    struct proc *curproc = myproc();
 
     int num = curproc->trap_frame->eax;
-    if (num > 0 && num < NELEM(syscalls) && syscalls[num])
-    {
+    if (num > 0 && num < (int)NELEM(syscalls) && syscalls[num]) {
         curproc->trap_frame->eax = syscalls[num]();
-    }
-    else
-    {
+    } else {
         cprintf("%d %s: unknown sys call %d\n",
-                curproc->pid, curproc->name, num);
+                curproc->pid,
+                curproc->name,
+                num);
         curproc->trap_frame->eax = -1;
     }
 }
